Avoid code duplication with logic related to `maxListeners`
