+++ lib/stdio/async.js
@@ -1,4 +1,3 @@
-import {addAbortListener} from 'node:events';
 import {createReadStream, createWriteStream} from 'node:fs';
 import {Buffer} from 'node:buffer';
 import {Readable, Writable} from 'node:stream';
@@ -6,7 +5,7 @@ import {handleInput} from './handle.js';
 import {pipeStreams} from './pipeline.js';
 import {TYPE_TO_MESSAGE} from './type.js';
 import {generatorToDuplexStream, pipeGenerator} from './generator.js';
-import {isStandardStream} from './utils.js';
+import {isStandardStream, incrementMaxListeners} from './utils.js';
 
 // Handle `input`, `inputFile`, `stdin`, `stdout` and `stderr` options, before spawning, in async mode
 export const handleInputAsync = options => handleInput(addPropertiesAsync, options, false);
@@ -72,25 +71,15 @@ const pipeStdioOption = (spawned, {type, value, direction, index}, inputStreamsG
 // Multiple processes might be piping from/to `process.std*` at the same time.
 // This is not necessarily an error and should not print a `maxListeners` warning.
 const setStandardStreamMaxListeners = (stream, {signal}) => {
-	if (!isStandardStream(stream)) {
-		return;
+	if (isStandardStream(stream)) {
+		incrementMaxListeners(stream, MAX_LISTENERS_INCREMENT, signal);
 	}
-
-	const maxListeners = stream.getMaxListeners();
-	if (maxListeners === 0 || maxListeners === Number.POSITIVE_INFINITY) {
-		return;
-	}
-
-	stream.setMaxListeners(maxListeners + maxListenersIncrement);
-	addAbortListener(signal, () => {
-		stream.setMaxListeners(stream.getMaxListeners() - maxListenersIncrement);
-	});
 };
 
 // `source.pipe(destination)` adds at most 1 listener for each event.
 // If `stdin` option is an array, the values might be combined with `merge-streams`.
 // That library also listens for `source` end, which adds 1 more listener.
-const maxListenersIncrement = 2;
+const MAX_LISTENERS_INCREMENT = 2;
 
 // The stream error handling is performed by the piping logic above, which cannot be performed before process spawning.
 // If the process spawning fails (e.g. due to an invalid command), the streams need to be manually destroyed.
