+++ lib/stdio/utils.js
@@ -1,4 +1,5 @@
 import {Buffer} from 'node:buffer';
+import {addAbortListener} from 'node:events';
 import process from 'node:process';
 
 export const bufferToUint8Array = buffer => new Uint8Array(buffer.buffer, buffer.byteOffset, buffer.byteLength);
@@ -12,3 +13,15 @@ export const binaryToString = uint8ArrayOrBuffer => textDecoder.decode(uint8Arra
 export const isStandardStream = stream => STANDARD_STREAMS.includes(stream);
 export const STANDARD_STREAMS = [process.stdin, process.stdout, process.stderr];
 export const STANDARD_STREAMS_ALIASES = ['stdin', 'stdout', 'stderr'];
+
+export const incrementMaxListeners = (eventEmitter, maxListenersIncrement, signal) => {
+	const maxListeners = eventEmitter.getMaxListeners();
+	if (maxListeners === 0 || maxListeners === Number.POSITIVE_INFINITY) {
+		return;
+	}
+
+	eventEmitter.setMaxListeners(maxListeners + maxListenersIncrement);
+	addAbortListener(signal, () => {
+		eventEmitter.setMaxListeners(eventEmitter.getMaxListeners() - maxListenersIncrement);
+	});
+};
